README + Errors
